# openapi.yaml
openapi: 3.0.3
info:
  title: AI Aggregator API
  version: 0.1.0
  description: REST interface that proxies multimodal AI requests to registered providers.
servers:
  - url: http://localhost:8080
paths:
  /v1/chat:
    post:
      summary: Generate text content using the default provider.
      tags: [chat]
      parameters:
        - $ref: '#/components/parameters/StrategyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextReq'
      responses:
        '200':
          description: Generated text
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TextRespEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/image:
    post:
      summary: Generate an image from a text prompt.
      tags: [image]
      parameters:
        - $ref: '#/components/parameters/StrategyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageReq'
      responses:
        '200':
          description: Generated image metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageRespEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/video:
    post:
      summary: Generate a short-form video.
      tags: [video]
      parameters:
        - $ref: '#/components/parameters/StrategyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoReq'
      responses:
        '200':
          description: Generated video metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoRespEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/stt:
    post:
      summary: Run speech-to-text transcription.
      tags: [speech]
      parameters:
        - $ref: '#/components/parameters/StrategyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/STTReq'
      responses:
        '200':
          description: Transcript payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/STTRespEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/tts:
    post:
      summary: Run text-to-speech synthesis.
      tags: [speech]
      parameters:
        - $ref: '#/components/parameters/StrategyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TTSReq'
      responses:
        '200':
          description: Synthesized audio reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TTSRespEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/embeddings:
    post:
      summary: Generate vector embeddings for inputs.
      tags: [embeddings]
      parameters:
        - $ref: '#/components/parameters/StrategyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmbeddingsReq'
      responses:
        '200':
          description: Embedding vectors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmbeddingsRespEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/moderation:
    post:
      summary: Run safety moderation checks on input text.
      tags: [moderation]
      parameters:
        - $ref: '#/components/parameters/StrategyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModerationReq'
      responses:
        '200':
          description: Moderation verdict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModerationRespEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/providers:
    get:
      summary: List registered upstream providers.
      tags: [providers]
      responses:
        '200':
          description: Active providers list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProvidersRespEnvelope'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/register:
    post:
      summary: Register a new user and issue an access/refresh token pair.
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterReq'
      responses:
        '200':
          description: Token pair for the new account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRespEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/login:
    post:
      summary: Exchange user credentials for an access/refresh token pair.
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginReq'
      responses:
        '200':
          description: Token pair for the authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRespEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/refresh:
    post:
      summary: Rotate refresh tokens and mint a new access token.
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshReq'
      responses:
        '200':
          description: Rotated token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRespEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/logout:
    post:
      summary: Revoke the provided refresh token and its session.
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutReq'
      responses:
        '200':
          description: Logout confirmation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutRespEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /session/{provider}/message:
    post:
      summary: Send a session-scoped text message using the user's stored provider key.
      tags: [sessions]
      parameters:
        - $ref: '#/components/parameters/SessionProviderParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionMessageReq'
      responses:
        '200':
          description: Provider response plus refreshed session metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionTextRespEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /session/{provider}/image:
    post:
      summary: Generate an image while tracking the user's provider session state.
      tags: [sessions]
      parameters:
        - $ref: '#/components/parameters/SessionProviderParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionImageReq'
      responses:
        '200':
          description: Image result and updated session counters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionImageRespEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /session/{provider}/video:
    post:
      summary: Generate a video artifact while preserving the user's session usage.
      tags: [sessions]
      parameters:
        - $ref: '#/components/parameters/SessionProviderParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionVideoReq'
      responses:
        '200':
          description: Video result plus session metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionVideoRespEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /session/{provider}/audio:
    post:
      summary: Transcribe audio using the user's provider session.
      tags: [sessions]
      parameters:
        - $ref: '#/components/parameters/SessionProviderParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionAudioReq'
      responses:
        '200':
          description: Transcript plus refreshed session tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionAudioRespEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
components:
  parameters:
    StrategyParam:
      name: strategy
      in: query
      description: Optional routing strategy (first, fast, cheap, bestOf)
      required: false
      schema:
        type: string
        enum: [first, fast, cheap, bestOf]
    SessionProviderParam:
      name: provider
      in: path
      description: Target provider identifier tied to the user's session.
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResp'
    InternalError:
      description: Provider execution failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResp'
    Unavailable:
      description: No providers registered
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResp'
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResp'
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResp'
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResp'
  schemas:
    ErrorResp:
      type: object
      properties:
        error:
          type: string
        provider:
          type: string
      additionalProperties: false
    RegisterReq:
      type: object
      properties:
        email:
          type: string
          format: email
        display_name:
          type: string
        password:
          type: string
          format: password
        preferences:
          type: object
          additionalProperties: true
        timezone:
          type: string
        locale:
          type: string
      required: [email, display_name, password]
    LoginReq:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
      required: [email, password]
    RefreshReq:
      type: object
      properties:
        refresh_token:
          type: string
      required: [refresh_token]
    LogoutReq:
      type: object
      properties:
        refresh_token:
          type: string
      required: [refresh_token]
    TextReq:
      type: object
      properties:
        provider:
          type: string
          description: Optional explicit provider identifier to override routing.
        prompt:
          type: string
        max_tokens:
          type: integer
        temperature:
          type: number
          format: float
      required: [prompt]
    TextResp:
      type: object
      properties:
        content:
          type: string
      required: [content]
    ImageReq:
      type: object
      properties:
        provider:
          type: string
          description: Optional explicit provider identifier to override routing.
        prompt:
          type: string
        size:
          type: string
      required: [prompt]
    ImageResp:
      type: object
      properties:
        url:
          type: string
          format: uri
      required: [url]
    VideoReq:
      type: object
      properties:
        provider:
          type: string
          description: Optional explicit provider identifier to override routing.
        prompt:
          type: string
        duration_seconds:
          type: integer
      required: [prompt]
    VideoResp:
      type: object
      properties:
        url:
          type: string
          format: uri
      required: [url]
    STTReq:
      type: object
      properties:
        provider:
          type: string
          description: Optional explicit provider identifier to override routing.
        audio_url:
          type: string
          format: uri
        language:
          type: string
      required: [audio_url]
    STTResp:
      type: object
      properties:
        transcript:
          type: string
      required: [transcript]
    TTSReq:
      type: object
      properties:
        provider:
          type: string
          description: Optional explicit provider identifier to override routing.
        text:
          type: string
        voice:
          type: string
        language:
          type: string
      required: [text]
    TTSResp:
      type: object
      properties:
        audio_url:
          type: string
          format: uri
      required: [audio_url]
    EmbeddingsReq:
      type: object
      properties:
        provider:
          type: string
          description: Optional explicit provider identifier to override routing.
        inputs:
          type: array
          items:
            type: string
        model:
          type: string
      required: [inputs]
    EmbeddingsResp:
      type: object
      properties:
        vectors:
          type: array
          items:
            type: array
            items:
              type: number
              format: float
      required: [vectors]
    ModerationReq:
      type: object
      properties:
        provider:
          type: string
          description: Optional explicit provider identifier to override routing.
        input:
          type: string
      required: [input]
    ModerationResp:
      type: object
      properties:
        flagged:
          type: boolean
        reason:
          type: string
    SessionBase:
      type: object
      properties:
        provider_key:
          type: string
          description: Optional API key used to bootstrap the session if no secret is stored yet.
        session_metadata:
          type: object
          additionalProperties: true
        expires_at:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false
    SessionMessageReq:
      allOf:
        - $ref: '#/components/schemas/SessionBase'
        - type: object
          properties:
            prompt:
              type: string
            max_tokens:
              type: integer
            temperature:
              type: number
              format: float
          required: [prompt]
    SessionImageReq:
      allOf:
        - $ref: '#/components/schemas/SessionBase'
        - type: object
          properties:
            prompt:
              type: string
            size:
              type: string
          required: [prompt]
    SessionVideoReq:
      allOf:
        - $ref: '#/components/schemas/SessionBase'
        - type: object
          properties:
            prompt:
              type: string
            duration_seconds:
              type: integer
          required: [prompt]
    SessionAudioReq:
      allOf:
        - $ref: '#/components/schemas/SessionBase'
        - type: object
          properties:
            audio_url:
              type: string
              format: uri
            language:
              type: string
          required: [audio_url]
    SessionDetails:
      type: object
      properties:
        provider_name:
          type: string
        provider_key:
          type: string
        last_interaction:
          type: string
          format: date-time
        total_tokens_used:
          type: integer
        expires_at:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
          additionalProperties: true
      required: [provider_name, provider_key, last_interaction, total_tokens_used, metadata]
    ProvidersResp:
      type: object
      properties:
        providers:
          type: array
          items:
            type: string
        matrix:
          type: array
          items:
            $ref: '#/components/schemas/CapabilityMatrixEntry'
      required: [providers, matrix]
    CapabilityMatrixEntry:
      type: object
      properties:
        name:
          type: string
        capabilities:
          $ref: '#/components/schemas/Capabilities'
      required: [name, capabilities]
    Capabilities:
      type: object
      properties:
        text_generation:
          type: boolean
        image_generation:
          type: boolean
        video_generation:
          type: boolean
        speech_to_text:
          type: boolean
        text_to_speech:
          type: boolean
        embeddings:
          type: boolean
        moderation:
          type: boolean
        limits:
          $ref: '#/components/schemas/CapabilityLimits'
        attributes:
          $ref: '#/components/schemas/CapabilityAttributes'
    CapabilityLimits:
      type: object
      properties:
        max_text_tokens:
          type: integer
        max_image_resolution:
          type: string
        max_embedding_dimensions:
          type: integer
    CapabilityAttributes:
      type: object
      properties:
        cost_score:
          type: integer
        latency_score:
          type: integer
        quality_score:
          type: integer
        rate_limit_rps:
          type: number
    AuthTokens:
      type: object
      properties:
        access_token:
          type: string
        access_expires_in:
          type: string
          description: Go duration string (e.g. "15m0s").
        refresh_token:
          type: string
        refresh_expires_in:
          type: string
          description: Go duration string (e.g. "168h0m0s").
      required: [access_token, access_expires_in, refresh_token, refresh_expires_in]
    LogoutResp:
      type: object
      properties:
        status:
          type: string
      required: [status]
    TextRespEnvelope:
      type: object
      properties:
        provider:
          type: string
        data:
          $ref: '#/components/schemas/TextResp'
      required: [provider, data]
    ImageRespEnvelope:
      type: object
      properties:
        provider:
          type: string
        data:
          $ref: '#/components/schemas/ImageResp'
      required: [provider, data]
    VideoRespEnvelope:
      type: object
      properties:
        provider:
          type: string
        data:
          $ref: '#/components/schemas/VideoResp'
      required: [provider, data]
    STTRespEnvelope:
      type: object
      properties:
        provider:
          type: string
        data:
          $ref: '#/components/schemas/STTResp'
      required: [provider, data]
    TTSRespEnvelope:
      type: object
      properties:
        provider:
          type: string
        data:
          $ref: '#/components/schemas/TTSResp'
      required: [provider, data]
    EmbeddingsRespEnvelope:
      type: object
      properties:
        provider:
          type: string
        data:
          $ref: '#/components/schemas/EmbeddingsResp'
      required: [provider, data]
    ModerationRespEnvelope:
      type: object
      properties:
        provider:
          type: string
        data:
          $ref: '#/components/schemas/ModerationResp'
      required: [provider, data]
    SessionTextRespEnvelope:
      type: object
      properties:
        provider:
          type: string
        data:
          $ref: '#/components/schemas/TextResp'
        session:
          $ref: '#/components/schemas/SessionDetails'
      required: [provider, data, session]
    SessionImageRespEnvelope:
      type: object
      properties:
        provider:
          type: string
        data:
          $ref: '#/components/schemas/ImageResp'
        session:
          $ref: '#/components/schemas/SessionDetails'
      required: [provider, data, session]
    SessionVideoRespEnvelope:
      type: object
      properties:
        provider:
          type: string
        data:
          $ref: '#/components/schemas/VideoResp'
        session:
          $ref: '#/components/schemas/SessionDetails'
      required: [provider, data, session]
    SessionAudioRespEnvelope:
      type: object
      properties:
        provider:
          type: string
        data:
          $ref: '#/components/schemas/STTResp'
        session:
          $ref: '#/components/schemas/SessionDetails'
      required: [provider, data, session]
    ProvidersRespEnvelope:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ProvidersResp'
      required: [data]
    AuthRespEnvelope:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/AuthTokens'
      required: [data]
    LogoutRespEnvelope:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/LogoutResp'
      required: [data]